added high goal shooting to autonoomous, some fine tuning needed still

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ArcmToSiA.java
@@ -498,9 +498,9 @@ public class ArcmToSiA extends LinearOpMode {
             int distance = 200; //random distance for rn
             int fullDistance = distance + 50; //need to find out the distance from landing point to front of power shot goal 50 is just a placer
             if(fullDistance <= 210) {
-                int rpmSPerct = (fullDistance / 216.74) * 100;
+                 double rpmSPerct = (fullDistance / 216.74) * 100;
             } else {
-                int rpmLPerct = (fullDistance / 404) * 100;
+                double rpmLPerct = (fullDistance / 404) * 100;
             }
